Moved lambda service to its own folder
