+++ ScoutSuite/providers/aws/services/awslambda/service.py
